requirements.txt
dbt-core~=1.7.4
dbt-bigquery~=1.7.9
dbt-postgres~=1.7.17
SQLAlchemy
google
google-cloud-bigquery-storage
fastapi 
uvicorn 
psycopg2
python-dotenv